pyproject.toml
[project]
name = "bacteriadb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "polars-lts-cpu>=1.19.0",
    "pydantic>=2.10.3",
]

[dependency-groups]
dev = []
